requirements.txt
# ─── Django core ──────────────────────────────────────────────────────────────
Django==5.2.7
djangorestframework==3.16.1
djangorestframework-simplejwt[crypto]==5.3.1  # L-5: [crypto] enables token_blacklist
django-filter==24.3
django-cors-headers==4.9.0
django-allauth==65.12.1
google-auth==2.38.0
requests==2.32.3
drf-spectacular==0.28.0
python-dotenv==1.1.1

# ─── PostgreSQL ───────────────────────────────────────────────────────────────
# psycopg[binary] bundles libpq — no system package needed in the Docker image
psycopg[binary]==3.2.6

# ─── Redis / cache ────────────────────────────────────────────────────────────
redis==5.2.1
django-redis==5.4.0

# ─── Celery ───────────────────────────────────────────────────────────────────
celery[redis]==5.4.0
django-celery-beat==2.9.0

# ─── Server ───────────────────────────────────────────────────────────────────
gunicorn==23.0.0
whitenoise==6.9.0

# ─── Media / QR ───────────────────────────────────────────────────────────────
qrcode==8.0
Pillow==11.1.0

# ─── Dev / lint / test ────────────────────────────────────────────────────────
python-dateutil==2.9.0.post0  # ISO date parsing in security tests
black==25.9.0
isort==7.0.0
flake8==7.3.0
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0
factory-boy==3.3.1
Faker==33.1.0  # Realistic test data generation
